+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change log
 
+## Unreleased
+
+- Updated test for `PdfTransparencyTrait` trait.
+
 ## 2.0.3 - 2024-08-14
 
 - Added `PdfTransparencyTrait` trait and the corresponding tutorial.